Revert download-file changes (will add on a different PR)
